refactor: extract screen reader and navigation logic to custom hooks
- Extract screen reader announcement logic to useScreenReaderAnnouncements hook
- Extract round navigation handlers to useRoundNavigation hook
- Improve code organization and reusability
- Maintain all existing functionality and test coverage

Addresses complexity issues found in RoundTracker component
