+++ .codacy/logs/codacy-cli.log
@@ -108,3 +108,9 @@
 2025-07-07T08:09:33-07:00 [INFO] (codacy-cli-v2/cli-v2.go:27) Starting Codacy CLI version=1.0.0-main.339.sha.dcd97cf (dcd97cf) built at 2025-06-16T20:10:29Z
 2025-07-07T08:09:33-07:00 [INFO] (codacy-cli-v2/cmd/root.go:34) Executing CLI command full_command=[/home/doug/.cache/codacy/codacy-cli-v2/1.0.0-main.339.sha.dcd97cf/codacy-cli-v2 analyze src/components/combat/round-tracking/round-utils.ts --format sarif] args=[src/components/combat/round-tracking/round-utils.ts] command=analyze
 2025-07-07T08:09:34-07:00 [DEBUG] (codacy-cli-v2/tools/revive/reviveRunner.go:53) Running Revive command command=/home/doug/.cache/codacy/tools/revive@1.7.0/revive -formatter sarif src/components/combat/round-tracking/round-utils.ts
+2025-07-07T16:16:34-07:00 [INFO] (codacy-cli-v2/cli-v2.go:27) Starting Codacy CLI version=1.0.0-main.339.sha.dcd97cf (dcd97cf) built at 2025-06-16T20:10:29Z
+2025-07-07T16:16:34-07:00 [INFO] (codacy-cli-v2/cmd/root.go:34) Executing CLI command full_command=[/home/doug/.cache/codacy/codacy-cli-v2/1.0.0-main.339.sha.dcd97cf/codacy-cli-v2 analyze --format sarif] args=[] command=analyze
+2025-07-07T16:16:34-07:00 [DEBUG] (codacy-cli-v2/tools/pmdRunner.go:32) PMD binary path pmdBinary=/home/doug/.cache/codacy/tools/pmd@6.55.0/pmd-bin-6.55.0/bin/run.sh
+2025-07-07T16:16:34-07:00 [DEBUG] (codacy-cli-v2/tools/pmdRunner.go:98) Setting up Java environment javaHome=/home/doug/.cache/codacy/runtimes/jdk-17.0.10+7
+2025-07-07T16:16:34-07:00 [DEBUG] (codacy-cli-v2/tools/pmdRunner.go:144) Updated environment variables binDir=/home/doug/.cache/codacy/runtimes/jdk-17.0.10+7/bin javaBinary=/home/doug/.cache/codacy/runtimes/jdk-17.0.10+7/bin/java javaHome=JAVA_HOME=/home/doug/.cache/codacy/runtimes/jdk-17.0.10+7 path=PATH=/home/doug/.cache/codacy/runtimes/jdk-17.0.10+7/bin:/home/doug/.npm/_npx/4f3c4ba89ff2c191/node_modules/.bin:/home/doug/dev/dnd-tracker-next-js/node_modules/.bin:/home/doug/dev/node_modules/.bin:/home/doug/node_modules/.bin:/home/node_modules/.bin:/node_modules/.bin:/home/doug/.npm-global/lib/node_modules/npm/node_modules/@npmcli/run-script/lib/node-gyp-bin:/home/doug/bin:/home/doug/.npm-global/bin:/home/doug/bin:/home/doug/bin:/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin:/usr/games:/usr/local/games:/snap/bin:/snap/bin
+2025-07-07T16:16:34-07:00 [DEBUG] (codacy-cli-v2/tools/pmdRunner.go:160) Running PMD command command=/home/doug/.cache/codacy/tools/pmd@6.55.0/pmd-bin-6.55.0/bin/run.sh pmd -R .codacy/tools-configs/ruleset.xml -d /home/doug/dev/dnd-tracker-next-js -f sarif -r /tmp/codacy-analysis-1052911152/pmd.sarif
